improvements

+++ controllers/ContactsController.js
@@ -6,27 +6,39 @@ class ContactsController {
 
     init(app, models) {
         app.get("/contacts", isAuthenticated, (req, res) => this.index.call(this, req, res, models));
-        app.get("/contacts/:id", (req, res) => this.contact.call(this, req, res, models));
-        app.post("/contacts", (req, res) => this.insert.call(this, req, res, models));
-        app.put("/contacts/:id", (req, res) => this.edit.call(this, req, res, models));
-        app.delete("/contacts/:id", (req, res) => this.delete.call(this, req, res, models));
+        app.get("/contacts/:id", isAuthenticated, (req, res) => this.contact.call(this, req, res, models));
+        app.post("/contacts", isAuthenticated, (req, res) => this.insert.call(this, req, res, models));
+        app.put("/contacts/:id", isAuthenticated, (req, res) => this.edit.call(this, req, res, models));
+        app.delete("/contacts/:id", isAuthenticated, (req, res) => this.delete.call(this, req, res, models));
     }
 
-    async index(req, res, { Contacto }) {
-        const contactos = await Contacto.find({});
-        res.send(contactos);
+    async index(req, res, { User }) {
+        try {
+            const user = await User.findById(req.me._id, 'contacts').populate('contacts');
+            res.status(200).json(user.contacts);
+        } catch (err) {
+            res.status(400).json(err);
+        }
     }
 
-    async insert(req, res, { Contacto }) {
+    async insert(req, res, { Contacto, User }) {
         const { name, telefone } = req.body;
-        const novo = new Contacto({
+        const newContact = new Contacto({
             name: name,
             telefone: telefone,
         });
         try {
-            await novo.save();
-            res.set("Content-Type", "application/json");
-            res.status(200).send(novo);
+            const user = await User.findById(req.me._id);
+            if (user) {
+                await newContact.save();
+                user.contacts.push(newContact._id);
+                await user.save();
+                res.status(201).json(newContact);
+            } else {
+                res.status(400).json({
+                    error: 'User Not Found',
+                });
+            }
         } catch (error) {
             res.status(400).json(error);
         }
@@ -49,24 +61,52 @@ class ContactsController {
         }
     }
 
-    async edit(req, res, { Contacto }) {
+    async edit(req, res, { Contacto, User }) {
         const id = req.params.id;
         const to_edit = req.body;
         try {
-            const contacto = await Contacto.findByIdAndUpdate(id, to_edit, { new: true });
-            res.json(contacto);
+            const userContacts = await User.findById(req.me._id);
+            const itsMine = userContacts.contacts.findIndex((value) => {
+                return value == id;
+            });
+            if (itsMine >= 0) {
+                const contacto = await Contacto.findByIdAndUpdate(id, to_edit, { new: true });
+                if(contacto) {
+                    res.status(200).json(contacto);
+                } else {
+                    res.status(400).json({error: 'No contact found'});
+                }
+            } else {
+                res.status(400).send({
+                    error: 'No Permission to update this contact'
+                });
+            }
         } catch (error) {
-            res.send(error);
+            res.status(400).send(error);
         }
     }
 
-    async delete(req, res, { Contacto }) {
+    async delete(req, res, { Contacto, User }) {
         const id = req.params.id;
         try {
-            const contacto = await Contacto.deleteOne({ _id: id });
-            res.json(contacto);
+            const userContacts = await User.findById(req.me._id);
+            const itsMine = userContacts.contacts.findIndex((value) => {
+                return value == id;
+            });
+            if (itsMine >= 0) {
+                const contacto = await Contacto.deleteOne({ _id: id });
+                if(contacto.deletedCount > 0) {
+                    res.status(200).json(contacto);
+                } else {
+                    res.status(400).json({error: 'No contact found'});
+                }
+            } else {
+                res.status(400).send({
+                    error: 'No Permission to delete this contact'
+                });
+            }
         } catch (error) {
-            res.send(error);
+            res.status(400).send(error);
         }
     }
 }
